requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
email-validator
Flask-Mail
python-dotenv
psycopg2-binary
gunicorn